# Migrate to @swc-node/jest to make tests running faster