--- a/package.json
+++ b/package.json
@@ -15,6 +15,7 @@
   ],
   "license": "MIT",
   "devDependencies": {
+    "@swc-node/jest": "^0.3.3",
     "@types/jest": "^25.2.1",
     "@types/react": "^16.9.34",
     "dts-bundle": "^0.7.3",
@@ -25,7 +26,6 @@
     "prettier": "^2.0.4",
     "rimraf": "^3.0.2",
     "semver": "^7.3.0",
-    "ts-jest": "~25.3.1",
     "tslint": "^6.1.1",
     "tslint-config-prettier": "^1.18.0",
     "tslint-plugin-prettier": "^2.3.0",