## `@swc-node/jest`

```
Test Suites: 4 passed, 4 total
Tests:       114 passed, 114 total
Snapshots:   213 passed, 213 total
Time:        4.394s, estimated 8s

Test Suites: 1 passed, 1 total
Tests:       11 passed, 11 total
Snapshots:   11 passed, 11 total
Time:        2.511s, estimated 7s

```

## `ts-jest`

```
Test Suites: 4 passed, 4 total
Tests:       114 passed, 114 total
Snapshots:   213 passed, 213 total
Time:        9.721s

Test Suites: 1 passed, 1 total
Tests:       11 passed, 11 total
Snapshots:   11 passed, 11 total
Time:        7.691s
```