@peduarte 2020-11-06
@Brooooooklyn thanks for this. We'll stick with `ts-jest` for now 👍